[MOD] Rename page components to partials
