+++ src/pages/sequencer/partials/controls.js
