+++ src/pages/sequencer/partials/index.js
