+++ src/pages/sequencer/partials/pianoroll.js
