+++ src/pages/synth/partials/amplifier.js
