+++ src/pages/synth/partials/arpegiator.js
