+++ src/pages/synth/partials/effects.js
