+++ src/pages/synth/partials/index.js
