+++ src/pages/synth/partials/live.js
