+++ src/pages/synth/partials/more.js
