+++ src/pages/synth/partials/oscilator.js
